Seperated modals in to components to clean up code
